explain/pyproject.toml
[project]
name = "fianchetto-explain"
version = "1.0.0"
description = "Explainability service for Fianchetto Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-chess>=1.999",
    "stockfish>=3.28.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

